pyproject.toml
[tool.poetry]
name = "red-purple"
version = "0.1.0"
description = "Red-Purple: Offensive + Iterative Learning AI Agent Framework"
authors = ["YounessB1"]
packages = [{include = "source"}]

[tool.poetry.scripts]
red-purple = "source.main:main"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
litellm = ">=1.81.1"
python-dotenv = "^1.0.0"
gepa = "*"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
